src/modules/accounts/useCases/authenticateUser/AuthenticateUserUseCase.ts
import { compare } from 'bcryptjs';
import { sign } from 'jsonwebtoken';
import { inject, injectable } from 'tsyringe';

import { IUserRepository } from '@modules/accounts/repositories/IUserRepository';
import { AppError } from '@shared/errors/AppErros';

interface IRequest {
    email: string;
    password: string;
}

interface IResponse {
    user: {
        name: string;
        email: string;
    };
    token: string;
}

@injectable()
class AuthenticateUserUseCase {
    constructor(
        @inject('UserRepository')
        private userRepository: IUserRepository,
    ) {}
    async execute({ email, password }: IRequest): Promise<IResponse> {
        const user = await this.userRepository.findByEmail(email);

        if (!user) {
            throw new AppError('Email or password incorrect!');
        }

        const passwordMatch = await compare(password, user.password);

        if (!passwordMatch) {
            throw new AppError('Email or password incorrect!');
        }

        const token = sign({}, 'b4cc344d25a2efe540adbf2678e2304c', {
            subject: user.id,
            expiresIn: '1d',
        });

        const tokeReturn: IResponse = {
            user: {
                email: user.email,
                name: user.name,
            },
            token,
        };

        return tokeReturn;
    }
}

export { AuthenticateUserUseCase };
